Ignore grants
Ignore grants on objects in databases that SnowDDL does not manage.
Grants on objects in databases that are not in _config folder are ignored during the resolver process. This allows clones of SnowDDL-managed databases to be taken (and owned by a different role), without causing SnowDDL to process a large number of revokes against the cloned database.